requirements.txt
fastapi[all]>=0.116.1
openai-whisper>=20250625
torch>=2.8.0
uvicorn>=0.35.0